Matchers refactored (Fixes #707)
RegexMatcher for matching strings to regular expressions was added (do
not mistake it with RegexMatcher which was in version 0.5)
WithRegexMatcher was marked as deprecated as I'd like to encourage
people to use RegexMatcher in combination with another matchers (for
example WithLabelMatcher, WithTooltipMatcher, ...)
WithTooltipMatcher was added because I was missing this functionality in
RedDeer.
All recurences of WithRegexMatcher and WithRegexMatchers were hopefully
removed.