Performance improvement of WidgetLookup.activeWidget
This method is used by AbstractWidget and therefore by all reddeer widgets,
which are created in most integration tests. So, it is one of methods
performance of which is critical.

Previous implementation collected the entire list of matching SWT widgets,
iterating over all widgets in referenced composite and then picked up
an element by the given index. Very often reference composite is not set
so that iteration goes over all widgets of the window.

The suggested implementation does the same iteration but without creating lists
and interrupting the iteration as soon as the requested number of matches is found
(and in most usages the first match is requested).

Otherwise, the logic of the search does not changed. Tests of Reddeer pass.